DOCS: Add Refresh Dataset Documentation
This addes Refresh Dataset documentation at the cost of removing docs
for running a report. Swagger schema does not allow for multiple
descriptions for the same path and http method combo. We either need to
add separate routing, or come up with a more creative solution for
describing these endpoints.